Move coding-system tests inline
